update thread module

+++ inc/thread.h
@@ -28,7 +28,7 @@ typedef struct{
 }context;
 
 typedef struct{
-	uint32_t* context;
+	uint8_t* context;
 	uint8_t status;
 	uint8_t priority;
 	uint32_t run_time;
